requirements.txt
soundcloud-v2>=1.5.0
yt-dlp>=2024.0.0
rich>=13.0.0
click>=8.0.0
requests>=2.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
